[IRDA]: Fix 64-bit warnings. 1) Use proper size_t printf format specifier 2) Cast pointers properly when passing them to hashfind 3) Print pointers using proper printf format specifier    instead of using ugly casts.

+++ net/irda/irda_device.c
@@ -231,7 +231,7 @@ static void __irda_task_delete(struct irda_task *task)
 void irda_task_delete(struct irda_task *task)
 {
 	/* Unregister task */
-	hashbin_remove(tasks, (int) task, NULL);
+	hashbin_remove(tasks, (long) task, NULL);
 
 	__irda_task_delete(task);
 }
@@ -345,7 +345,7 @@ struct irda_task *irda_task_execute(void *instance,
 	init_timer(&task->timer);
 
 	/* Register task */
-	hashbin_insert(tasks, (irda_queue_t *) task, (int) task, NULL);
+	hashbin_insert(tasks, (irda_queue_t *) task, (long) task, NULL);
 
 	/* No time to waste, so lets get going! */
 	ret = irda_task_kick(task);
